Create Calculate to the Sun rays
The angle of incidence of the sun's rays varies in daily, monthly and annual temporal range depending on the shape and movements of the earth. With this example application, the principal components of climate components can be understood by analysing the angle of incidence of solar radiation and shadow length for the living space.